pyproject.toml
[project]
name = "music-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "librosa",
    "matplotlib",
    "midiutil",
    "mido",
    "numpy",
    "pygame-ce>=2.5.6",
    "scipy",
    "taskipy>=1.14.1",
]

[dependency-groups]
dev = [
    "black>=25.12.0",
    "isort>=7.0.0",
    "ruff>=0.14.10",
]


#######################################################
# LINTERS - RUFF | Isort | Black
#######################################################
[tool.ruff]
line-length = 120
lint.select = ["E", "F", "I"]
# select = ["A", "B", "C", "D", "E", "F", "I", "N", "PL", "RUF", "S", "T", "W"]
lint.fixable = ["ALL"]
lint.ignore = ["D203", "D213", "T201", "PLE1205"]
lint.per-file-ignores = {}
# Same as Black.

# Allow unused variables when underscore-prefixed.
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
exclude = [
    "tests/*",
    ".venv/*",
    "build",
    "dist",
    "node_modules"
]

# [tool.ruff.lint.pydocstyle]
# convertion="google"


[tool.black]
line-length = 120
extend-exclude='''
.temp-*
| venv
| .venv
'''

#######################################################
# Poetry Scripts & Taskipy
#######################################################
[tool.poetry.scripts]
syncro_git = "src.fix:syncro_git"

[tool.taskipy.tasks]
pre_lint = { cmd = "echo 'Black' && black .", help = "apply the black linter"}
lint = { cmd = "ruff check . --fix", help = "apply ruff linter"}
tree = "uv tree > .pip.tree"
